% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_mx_dataset.R
\name{check_data_values}
\alias{check_data_values}
\title{Internal function to check data values in validation step}
\usage{
check_data_values(data_values)
}
\value{
boolean TRUE if passes
}
\description{
Internal function to check data values in validation step
}
