% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otsu_helpers.R
\name{check_threshold}
\alias{check_threshold}
\title{Internal function to calculate misclassification error}
\usage{
check_threshold(vec, thr1, thr2)
}
\arguments{
\item{vec}{vector of values to compare}

\item{thr1}{first threshold used to compare values}

\item{thr2}{second threshold used to compare values}
}
\value{
decimal value of misclassification error
}
\description{
Internal function to calculate misclassification error
}
