% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combat_helpers.R
\name{delta_conv}
\alias{delta_conv}
\title{Internal function to check convergence of delta terms}
\usage{
delta_conv(batch_chan, delta_stars, slide_var)
}
\arguments{
\item{batch_chan, delta_stars, slide_var}{Internal objects created/used in \code{run_combat}}
}
\value{
MAE between previous and updated delta values
}
\description{
Internal function to check convergence of delta terms
}
