% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combat_helpers.R
\name{gamma_conv}
\alias{gamma_conv}
\title{Internal function to check convergence of gamma terms}
\usage{
gamma_conv(batch_chan, gamma_stars, slide_var)
}
\arguments{
\item{batch_chan, gamma_stars, slide_var}{Internal objects created/used in \code{run_combat}}
}
\value{
MAE between previous and updated gamma values
}
\description{
Internal function to check convergence of gamma terms
}
