% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{get_ad_test_stats}
\alias{get_ad_test_stats}
\title{Internal function to generate Anderson-Darling test statistics}
\usage{
get_ad_test_stats(data_table, mx_data, n_bins = 100)
}
\arguments{
\item{data_table}{mx_data table to provide statistics}

\item{mx_data}{\code{mx_dataset} object to use to calc statistics}

\item{n_bins}{number of bins for density calculation (default=100). Large values of \code{n_bins} will slow computation considerably.}
}
\value{
\code{data.table} with AD test results
}
\description{
Internal function to generate Anderson-Darling test statistics
}
