% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_long_data.R
\name{get_long_data}
\alias{get_long_data}
\title{Internal function to create long data object from \code{mx_dataset} object}
\usage{
get_long_data(mx_data)
}
\arguments{
\item{mx_data}{\code{mx_dataset} object}
}
\value{
\code{data.frame} in long format
}
\description{
Internal function to create long data object from \code{mx_dataset} object
}
