% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_otsu_tab.R
\name{get_otsu_tab}
\alias{get_otsu_tab}
\title{Title}
\usage{
get_otsu_tab(tdat, cols, slide, table, threshold, use_np_array)
}
\arguments{
\item{tdat}{data in the \code{mx_dataset} object used to calculate Otsu thresholds}

\item{cols}{columns of markers in the  \code{mx_dataset} object used to calculate Otsu threshold}

\item{slide}{column identifying different slides in the  \code{mx_dataset} object used to calculate Otsu threshold}

\item{table}{dataset in \code{mx_data} used to compute metrics. Options include: c("raw","normalized","both"), e.g. a y-axis parameter.}

\item{threshold}{thresolding function as defined using \code{otsu_misclass()}}

\item{use_np_array}{boolean indicating if np_arrays are necessary}
}
\value{
list of calculated Otsu thresholds
}
\description{
Title
}
