% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{get_umap_cluster_results}
\alias{get_umap_cluster_results}
\title{Internal function to generate UMAP clustering results}
\usage{
get_umap_cluster_results(table, mx_data)
}
\arguments{
\item{table}{table in \code{mx_data} to provide clustering results}

\item{mx_data}{\code{mx_dataset} object to use to generate clustering results results}
}
\value{
\code{data.frame} with UMAP clustering results
}
\description{
Internal function to generate UMAP clustering results
}
