% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_umap_data.R
\name{get_umap_data}
\alias{get_umap_data}
\title{Internal function to setup dataset for UMAP algorithm}
\usage{
get_umap_data(mx_data, table, marker_list, metadata_cols)
}
\arguments{
\item{mx_data}{\code{mx_dataset} object used to compute UMAP dimensions}

\item{table}{dataset in \code{mx_data} used to compute metrics. Options include: c("raw","normalized","both"), e.g. a y-axis parameter.}

\item{marker_list}{list of markers in the \code{mx_dataset} object to use for UMAP algorithm}

\item{metadata_cols}{other identifiers of the input \code{data} (default=NULL). This must be a vector of column name(s) in the \code{mx_dataset} object}
}
\value{
data.frame for umap analysis
}
\description{
Internal function to setup dataset for UMAP algorithm
}
