% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_umap_reduction.R
\name{get_umap_reduction}
\alias{get_umap_reduction}
\title{Internal function to run UMAP algorithm}
\usage{
get_umap_reduction(udata, marker_list, downsample_pct)
}
\arguments{
\item{udata}{dataset to use for UMAP algorithm, generated with \code{get_umap_data()}}

\item{marker_list}{list of markers in the \code{mx_dataset} object to use for UMAP algorithm}

\item{downsample_pct}{double, optional percentage (0, 1] of sample rows to include when running UMAP algorithm. (default=1)}
}
\value{
data.frame
}
\description{
Internal function to run UMAP algorithm
}
