% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_data.R
\name{get_var_data}
\alias{get_var_data}
\title{Internal function to setup variance proportions table}
\usage{
get_var_data(mx_data, table, mod, m)
}
\arguments{
\item{mx_data}{\code{mx_dataset} object used to compute UMAP dimensions}

\item{table}{dataset in \code{mx_data} used to compute metrics. Options include: c("raw","normalized","both"), e.g. a y-axis parameter.}

\item{mod}{\code{lme4::lmer()} model run using \code{get_var_proportions()}}

\item{m}{marker used in \code{lme4} random effects model}
}
\value{
\code{data.frame}
}
\description{
Internal function to setup variance proportions table
}
