% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_sample.R
\docType{data}
\name{mx_sample}
\alias{mx_sample}
\title{Sample multiplexed dataset for \code{mxnorm}.}
\format{
A data frame with 3000 rows and 6 variables:
\describe{
\item{slide_id}{slide identifier}
\item{image_id}{image identifier}
\item{marker1_vals}{simulated marker 1 values}
\item{marker2_vals}{simulated marker 2 values}
\item{marker3_vals}{simulated marker 3 values}
\item{metadata1_vals}{simulated metadata 1 values}
...
}
}
\usage{
mx_sample
}
\description{
A simulated multiplexed dataset containing:
4 slides, with 3 images each, and 3 markers, and 1 metadata column.
}
\keyword{datasets}
