% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_mx_combat.R
\name{normalize_mx_combat}
\alias{normalize_mx_combat}
\title{Internal function to normalize mx_dataset using ComBat}
\usage{
normalize_mx_combat(mx_data, remove_zeroes = FALSE, tol = 1e-04, ...)
}
\arguments{
\item{mx_data}{\code{mx_dataset} object to normalize}

\item{remove_zeroes, tol, ...}{optional additional arguments for \code{normalize_mx_combat}}
}
\value{
\code{mx_dataset} object with normalized data with updated attribute \code{norm_data} (data.frame) and new attribute \code{method} (character)
}
\description{
Internal function to normalize mx_dataset using ComBat
}
