% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_mx_registration.R
\name{normalize_mx_registration}
\alias{normalize_mx_registration}
\title{Internal function to normalize mx_dataset using functional data registration}
\usage{
normalize_mx_registration(
  mx_data,
  len = 512,
  weighted = TRUE,
  offset = 1e-04,
  fdobj_norder = 4,
  fdobj_nbasis = 21,
  w_norder = 2,
  w_nbasis = 2,
  ...
)
}
\arguments{
\item{mx_data}{\code{mx_dataset} object to normalize}

\item{len, weighted, offset, fdobj_norder, fdobj_nbasis, w_norder, w_nbasis, ...}{optional additional arguments for \code{normalize_mx_registration}}
}
\value{
\code{mx_dataset} object with normalized data with updated attribute \code{norm_data} (data.frame) and new attribute \code{method} (character)
}
\description{
Internal function to normalize mx_dataset using functional data registration
}
