% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otsu_mx_agreement.R
\name{otsu_mx_agreement}
\alias{otsu_mx_agreement}
\title{Internal function to run agreement score analysis on otsu table}
\usage{
otsu_mx_agreement(mx_data, table)
}
\arguments{
\item{mx_data}{\code{mx_dataset} object used to compute Otsu agreement scores}

\item{table}{dataset in \code{mx_data} used to compute metrics. Options include: c("raw","normalized","both"), e.g. a y-axis parameter.}
}
\value{
\code{mx_dataset} object with agreement results
}
\description{
Internal function to run agreement score analysis on otsu table
}
