% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otsu_mx_dataset.R
\name{otsu_mx_dataset}
\alias{otsu_mx_dataset}
\title{Internal function to generate table for otsu misclassification analysis}
\usage{
otsu_mx_dataset(mx_data, table, threshold)
}
\arguments{
\item{mx_data}{\code{mx_dataset} object used to compute Otsu discordance scores}

\item{table}{dataset in \code{mx_data} used to compute metrics. Options include: c("raw","normalized","both"), e.g. a y-axis parameter.}

\item{threshold}{thresolding function as defined using \code{run_otsu_discordance()}}
}
\value{
\code{mx_dataset} object with \code{otsu_data} attribute added and misclassification results
}
\description{
Internal function to generate table for otsu misclassification analysis
}
