% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otsu_mx_discordance.R
\name{otsu_mx_discordance}
\alias{otsu_mx_discordance}
\title{Internal function to run discordance score analysis on otsu table}
\usage{
otsu_mx_discordance(mx_data, table)
}
\arguments{
\item{mx_data}{\code{mx_dataset} object used to compute Otsu discordance scores}

\item{table}{dataset in \code{mx_data} used to compute metrics. Options include: c("raw","normalized","both"), e.g. a y-axis parameter.}
}
\value{
\code{mx_dataset} object with discordance results
}
\description{
Internal function to run discordance score analysis on otsu table
}
