% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otsu_helpers.R
\name{otsu_mx_error}
\alias{otsu_mx_error}
\title{Internal helper function to compute and return Otsu table}
\usage{
otsu_mx_error(tdat, table, otsu_data, slides, cols)
}
\arguments{
\item{tdat}{table of raw or normalized values}

\item{table}{dataset in \code{mx_data} used to compute metrics. Options include: c("raw","normalized","both"), e.g. a y-axis parameter.}

\item{otsu_data}{otsu dataset created using \code{otsu_mx_dataset()}}

\item{slides}{vector of slides in the dataset}

\item{cols}{vector of marker columns}
}
\value{
\code{data.frame} with Otsu misclassification metrics
}
\description{
Internal helper function to compute and return Otsu table
}
