% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otsu_mx_misclassification.R
\name{otsu_mx_misclassification}
\alias{otsu_mx_misclassification}
\title{Internal function to run misclassification analysis on otsu table}
\usage{
otsu_mx_misclassification(mx_data, table)
}
\arguments{
\item{mx_data}{\code{mx_dataset} object used to compute Otsu misclassification metrics}

\item{table}{dataset in \code{mx_data} used to compute metrics. Options include: c("raw","normalized","both"), e.g. a y-axis parameter.}
}
\value{
\code{mx_dataset} object with misclassification results
}
\description{
Internal function to run misclassification analysis on otsu table
}
