% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mx_agreement.R
\name{plot_mx_agreement}
\alias{plot_mx_agreement}
\title{Visualize Otsu agreement scores by marker and slide}
\usage{
plot_mx_agreement(mx_data)
}
\arguments{
\item{mx_data}{\code{mx_dataset} object that been used with \code{run_otsu_agreement()} to compute Otsu agreement scores. Note that the table attribute must be set when running \code{run_otsu_agreement()}.}
}
\value{
\code{ggplot2} object with Otsu agreement scores plot
}
\description{
Visualize Otsu agreement scores by marker and slide
}
\examples{
mx_data = mx_dataset(mxnorm::mx_sample, "slide_id", "image_id",
  c("marker1_vals","marker2_vals","marker3_vals"),
  c("metadata1_vals"))
mx_data = mx_normalize(mx_data, transform="log10",method="None")
mx_data = run_otsu_agreement(mx_data, table="normalized")
plot_mx_agreement(mx_data)
}
