% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_mx_dataset.R
\name{scale_mx_dataset}
\alias{scale_mx_dataset}
\title{Internal function to scale mx_dataset}
\usage{
scale_mx_dataset(mx_data, scale)
}
\arguments{
\item{mx_data}{\code{mx_dataset} object to normalize}

\item{scale}{scale transformation to perform on the input data. Options include: c("None", "log10", "mean_divide","log10_mean_divide")}
}
\value{
\code{mx_dataset} object with scaled data
}
\description{
Internal function to scale mx_dataset
}
