% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\title{Setup S3 method for \code{summary}}
\usage{
summary(x)
}
\arguments{
\item{x}{used to define S3 method (later assumed to be \code{mx_dataset} object)}
}
\description{
Setup S3 method for \code{summary}
}
