% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_mx_dataset.R
\name{transform_mx_dataset}
\alias{transform_mx_dataset}
\title{Internal function to transform mx_dataset}
\usage{
transform_mx_dataset(mx_data, transform)
}
\arguments{
\item{mx_data}{\code{mx_dataset} object to normalize}

\item{transform}{transformation to perform on the input data. Options include: c("None", "log10", "mean_divide","log10_mean_divide")}
}
\value{
\code{mx_dataset} object with transformed data, added attributes: \code{norm_data} (data.frame) and \code{transform} (character)
}
\description{
Internal function to transform mx_dataset
}
