% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combat_helpers.R
\name{update_delta}
\alias{update_delta}
\title{Update ComBat delta (variance) parameters each iteration of the algorithm}
\usage{
update_delta(batch_chan, beta_c, omega_c, channel, slide_var)
}
\arguments{
\item{batch_chan, beta_c, omega_c, channel, slide_var}{Internal objects created/used in \code{run_combat}}
}
\value{
Updated delta values
}
\description{
Update ComBat delta (variance) parameters each iteration of the algorithm
}
