% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combat_helpers.R
\name{update_gamma}
\alias{update_gamma}
\title{Update ComBat gamma (mean) parameters each iteration of the algorithm}
\usage{
update_gamma(batch_chan, gamma_c, tau_c, channel, slide_var)
}
\arguments{
\item{batch_chan, gamma_c, tau_c, channel, slide_var}{Internal objects created/used in \code{run_combat}}
}
\value{
Updated gamma values
}
\description{
Update ComBat gamma (mean) parameters each iteration of the algorithm
}
