% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_mx_dataset.R
\name{validate_mx_dataset}
\alias{validate_mx_dataset}
\title{Internal function to validate mx_dataset object}
\usage{
validate_mx_dataset(x)
}
\arguments{
\item{x}{mx_dataset object to validate}
}
\value{
mx_dataset object
}
\description{
Internal function to validate mx_dataset object
}
