% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_threshold_override.R
\name{validate_threshold_override}
\alias{validate_threshold_override}
\title{Internal function to validate user defined thresholding algorithm}
\usage{
validate_threshold_override(threshold_override, ...)
}
\arguments{
\item{threshold_override}{optional user-defined function or alternate thresholding algorithm adaptable from Python skimage module \code{filters} (Note: not all algorithms in \code{filters} adapted). Options include supplying a function or any of the following: c("isodata", "li", "mean", "otsu", "triangle","yen"). More detail available here:https://scikit-image.org/docs/dev/api/skimage.filters.html. If using a user-defined function, it must include a \code{thold_data} parameter.}

\item{...}{optional additional arguments for Otsu agreement functions}
}
\value{
threshold_override function if it passes
}
\description{
Internal function to validate user defined thresholding algorithm
}
