% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration_helpers.R
\name{weighted.mean.fd}
\alias{weighted.mean.fd}
\title{Calculate weighted mean of fda object}
\usage{
\method{weighted.mean}{fd}(x, w, ...)
}
\arguments{
\item{x, w, ...}{Internal objects created/used in \code{run_registration}}
}
\value{
Weighted mean of fda object
}
\description{
Calculate weighted mean of fda object
}
