% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otsu_misclass.R
\name{otsu_misclass}
\alias{otsu_misclass}
\title{Title}
\usage{
otsu_misclass(
  mx_data,
  table,
  metadata_cols = NULL,
  threshold_override = NULL,
  plot_out = TRUE
)
}
\arguments{
\item{mx_data}{\code{mx_dataset} object used to compute Otsu misclassification metrics}

\item{table}{dataset in \code{mx_data} used to compute metrics. Options include: c("raw","normalized","both"), e.g. a y-axis parameter.}

\item{metadata_cols}{column name(s) in \code{mx_data} used to generate stratified results (default=NULL), e.g. a facet parameter.}

\item{threshold_override}{optional user-defined function or alternate thresholding algorithm in Python skimage module \code{filters}. Options include supplying a function or any of the following: c("isodata", "li", "local", "mean", "minimum", "multiotsu", "niblack", "otsu", "sauvola", "triangle"). More detail available here:https://scikit-image.org/docs/dev/api/skimage.filters.html}

\item{plot_out}{boolean to generate Otsu misclassification plot (default=FALSE)}
}
\value{
\code{mx_dataset} object with analysis results of Otsu misclassification in \code{otsu_data} table
}
\description{
Title
}
