% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otsu_mx_dataset.R
\name{otsu_mx_dataset}
\alias{otsu_mx_dataset}
\title{Internal function to run otsu misclassification analysis}
\usage{
otsu_mx_dataset(mx_data, table, metadata_cols, threshold)
}
\arguments{
\item{mx_data}{\code{mx_dataset} object used to compute Otsu misclassification metrics}

\item{table}{dataset in \code{mx_data} used to compute metrics. Options include: c("raw","normalized","both"), e.g. a y-axis parameter.}

\item{metadata_cols}{column name(s) in \code{mx_data} used to generate stratified results (default=NULL), e.g. a facet parameter.}

\item{threshold}{thresolding function as defined using \code{otsu_misclass()}}
}
\value{
\code{mx_dataset} object with \code{otsu_data} attribute added and misclassification results
}
\description{
Internal function to run otsu misclassification analysis
}
