% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.mx_dataset}
\alias{print.summary.mx_dataset}
\title{Extension of \code{print} S3 method to print \code{summary.mx_dataset} objects}
\usage{
\method{print}{summary.mx_dataset}(x, ...)
}
\arguments{
\item{x}{\code{summary.mx_dataset} object to summarize}

\item{...}{option for additional params given S3 logic}
}
\description{
Extension of \code{print} S3 method to print \code{summary.mx_dataset} objects
}
\examples{
mx_data = mx_dataset(mxnorm::mx_sample, "slide_id", "image_id",
  c("marker1_vals","marker2_vals","marker3_vals"),
  c("metadata1_vals"))
print(summary(mx_data))
}
