% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.mx_dataset}
\alias{summary.mx_dataset}
\title{Extension of \code{summary} S3 method to summarize \code{mx_dataset} objects}
\usage{
\method{summary}{mx_dataset}(object, ...)
}
\arguments{
\item{object}{\code{mx_dataset} object to summarize}

\item{...}{option for additional params given S3 logic}
}
\value{
\code{summary.mx_dataset} object
}
\description{
Extension of \code{summary} S3 method to summarize \code{mx_dataset} objects
}
\examples{
mx_data = mx_dataset(mxnorm::mx_sample, "slide_id", "image_id",
  c("marker1_vals","marker2_vals","marker3_vals"),
  c("metadata1_vals"))
summary(mx_data)
}
