% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.mx_dataset}
\alias{summary.mx_dataset}
\title{Extension of \code{summary} S3 method to summarize \code{mx_dataset} objects}
\usage{
\method{summary}{mx_dataset}(x)
}
\arguments{
\item{x}{\code{mx_dataset} object to summarize}
}
\value{
TBD
}
\description{
Extension of \code{summary} S3 method to summarize \code{mx_dataset} objects
}
\examples{
summary(mxnorm::mx_sample)
}
